Moved opencv support to separate target
